renamed enum, constructor adjustment
